xilinx-zynq: add source code for binary bootloader
RallyID: US109729

In order to comply with GPLv2 license, it is need to integrate
the corresponding source code of any binary file.
The binary bootloader is comprise of below components:
- FSBL(first stage bootloader)
- FPGA design
- u-boot
Therefore, the three components source code are integrated into
wrlinux production.

In addtional, add a README to describe the bootloader in detail.

Signed-off-by: Meng Li <Meng.Li@windriver.com>