amd-zynq: include zynq-ostree.conf in bsp-pkgs.conf
Issue: LINUXEXEC-33568

The ostree feature is enabled through the ostree template, but the
template parsing occurs much later than the parsing of the machine
conf amd-zynq.conf. As a result, when the machine conf is parsed,
"ostree" has not yet been added into DISTRO_FEATURES. Therefore,
we can't conditionally include the ostree conf by checking the
DISTRO_FEATURES in machine conf. Since the bsp-pkgs.conf is also
based on the template mechanism, we can move the check to bsp-pkgs.conf
to address this issue.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>