amd-zynq: use zImage as the only kernel image type
Issue: LIN1024-5821

Once we assign a value to WRL_ZYNQ_BOOTBIN_TARGET, the config files in
meta-xilinx will be involved. And the value of KERNEL_IMAGETYPE will be
set to "uImage" because "uImage" is added to IMAGE_BOOT_FILES in
    meta-xilinx-core/conf/machine/zynq-generic.conf

In WRLinux, we only provide zImage for amd-zynq, so fix KERNEL_IMAGETYPE
to "zImage" and clean IMAGE_BOOT_FILES to stop "uImage" from getting
involved.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>